Add files via upload
This python code is very useful to demonstrate the Monty Hall paradox.
The Monty Hall paradox occurred because of a challenge of a famous American TV show in the 60's and 70's, in which the contestant had to choose between 3 closed doors. Behind two of these doors was a goat and on the remaining door the prize was a car. The participant after choosing one door watched as the presenter opened one of the two remaining doors and showed a goat (the presenter always opens the door with the goat), then the participant was presented with another question "Do you wish to choose the other door or keep your initial choice?". Here is where the paradox is found, faced with this choice the ordinary participant would think that he/she does not have to choose the other door since he/she has a 50% chance of getting it right. But later, Marylin vos Savant, one of the people with the highest IQ at the time, would say that you have a 66% chance of winning if you change doors, which would be twice as likely if you decided to keep your initial choice.
This Python script asks you how many times you want to repeat the experience and at the end of it the number of times you win by changing, keeping the initial choice and their respective percentages.